Refactoring Config.java. Method _setupFakeTestingContext() was moved to ConfigTestHelper.java
