+++ src/com/dotmarketing/util/Config.java
@@ -56,8 +56,8 @@ public class Config {
 	private static Date lastRefreshTime = new Date ();
 	private static PropertiesConfiguration props = null;
 	private static ClassLoader classLoader = null;
-    private static URL dotmarketingPropertiesUrl = null;
-    private static URL clusterPropertiesUrl = null;
+    protected static URL dotmarketingPropertiesUrl = null;
+    protected static URL clusterPropertiesUrl = null;
     private static int prevInterval = Integer.MIN_VALUE;
 
     private static final String syncMe = "esSync";
@@ -499,33 +499,4 @@ public class Config {
 		lastRefreshTime = new Date(0);
 	}
 
-	/**
-	 * This method will set up a dummy ServletContext needed for testing. The main purpose here
-	 * is to be able to run the integration tests without the web app container i.e. Tomcat.
-	 *
-	 * @throws Exception
-     */
-	@VisibleForTesting
-	public static void _setupFakeTestingContext() throws Exception{
-		// if we need a fake ServletContext
-		if(CONTEXT ==null){
-			ServletContext context = Mockito.mock(ServletContext.class);
-			Mockito.when(context.getInitParameter("company_id")).thenReturn("dotcms.org");
-
-			final String topPath= Files.createTempDir().getCanonicalPath();
-			Mockito.when(context.getRealPath(Matchers.anyString())).thenAnswer(new Answer<String>() {
-				@Override
-				public String answer(InvocationOnMock invocation) throws Throwable {
-					String path = (String) invocation.getArguments()[0];
-					path = topPath + path.replaceAll("/", File.separator);
-
-					return path;
-				}
-			});
-			Config.CONTEXT = context;
-		}
-
-		dotmarketingPropertiesUrl = new File("test-resources/it-dotmarketing-config.properties").toURI().toURL();
-		clusterPropertiesUrl = new File("test-resources/it-dotcms-config-cluster.properties").toURI().toURL();
-	}
 }
